Fixing Fix #25: Replace DEFAULT_MAT_VOLT_FACTOR by command line
Fix #25 changed the value of DEFAULT_MAT_VOLT_FACTOR in extract/tools.py
to 1. This breaks the tutorials, because the simulations used in them
have to be up-scaled by a factor of 100.

As a solution, we added a new option to `css-extract`, namely
`--matfile-scale-factor`. This will be used in the tutorials from now
on.